refactor: split bnr and policy views into different files.
