+++ enterprise_access/apps/api/v1/views/__init__.py
@@ -1,0 +1,11 @@
+"""
+Top-level views module for convenience of maintaining
+existing imports of browse and request AND access policy views.
+"""
+from .browse_and_request import (
+    CouponCodeRequestViewSet,
+    LicenseRequestViewSet,
+    SubsidyRequestCustomerConfigurationViewSet,
+    SubsidyRequestViewSet
+)
+from .subsidy_access_policy import SubsidyAccessPolicyCRUDViewset, SubsidyAccessPolicyRedeemViewset
