+++ pytest.local.ini
@@ -2,8 +2,7 @@
 # when running pytest locally, for example:
 # pytest -x enterprise_access/apps/events/  --cov=enterprise_access.apps.events -c pytest.local.ini
 [pytest]
-DJANGO_SETTINGS_MODULE = enterprise_access.settings.test
-addopts = --cov-report term-missing --cov-report xml -W ignore
+addopts = --cov-report term-missing --cov-report xml -W ignore --ds=enterprise_access.settings.test
 norecursedirs = .* docs requirements site-packages
 
 # Filter depr warnings coming from packages that we can't control.
